% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_event_levels.R
\name{log_event_level}
\alias{log_event_level}
\title{Log Event Levels}
\usage{
log_event_level(class, level_number)
}
\arguments{
\item{class}{S3 class of the event; <character>}

\item{level_number}{level number; <integer>}
}
\value{
log event constructor; <function>
}
\description{
Log Event Levels
}
