% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{with_limits}
\alias{with_limits}
\title{Set limits to a logger}
\usage{
with_limits(logger, lower = NULL, upper = NULL)
}
\arguments{
\item{logger}{logger; <logger>}

\item{lower}{lower limit; <numeric> | <log_event_level>}

\item{upper}{uppwer limit; <numeric> | <log_event_level>}
}
\value{
logger; <logger>
}
\description{
Log events with level number smaller than the lower limit or
higher than the upper limit will be ignored by the logger.
}
\examples{
log_this <- logger() \%>\%
    with_limits(lower = 20,
                upper = logthis::WARNING)
}
