% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{with_receivers}
\alias{with_receivers}
\title{Add receivers to a logger}
\usage{
with_receivers(logger, ..., append = TRUE)
}
\arguments{
\item{logger}{a logger; <logger>}

\item{...}{receivers; <log_receiver>}

\item{append}{to append to existing receivers, or to overwrite; <logical>}
}
\value{
logger; <logger>
}
\description{
Add receivers to a logger
}
\examples{

log_this <- logger() %>%
    with_receivers(to_identity(),
                   to_console())

}
